第三方广播修改为静态广播
